algo/requirements.txt
django-extensions==3.2.1
djangorestframework==3.13.1
mysqlclient
pandas
numpy
openpyxl
opencv-python-headless
pillow
boto3
django-storages==1.13.1
tensorflow